Fix volumes test for native docker
The test was checking that a container's name was in the VolumesFrom
field. Docker allows for either id or name to be specified and reflects
that choice in the docker inspect. With native-docker, we now use the
id instead of the name, which is stored in the instance's externalId
field. So, we just need to change the test to use the externalId.