Read websocket stream until whole msg is received
The tests for container execing were assuming that the exact response
they wanted were always coming back in a single message, but there is
no guarantee to the message size and now it comes back in chunks. So,
keep reading from the websocket until the whole message is received (or
we eventually time out).