Fix native docker lxc, memory, and network tests
Remove lxc_conf from the native docker test since the driver in Docker
1.7 doesn't support it.

Pass host_config into create_container instead of into start since that
was deprecated in the Docker API and it was cauing memory settings to
be stomped on.

Added busybox to the list of images to pull since we now use it in a
test.