Fix tests wrt latest auth changes
Auth-service now returns 403 for forbidden users with valid creds.
On testlogin, all settings are used now to check whether admin can
still be found. The current test for groupSearchDomain was setting
user search domain in a way that admin wasn't a part of it, making
the test fail when run against latest code. This commit fixes that